+++ .scripts/classes/Compile.js
@@ -46,9 +46,19 @@ export class Compile extends CompileStage {
 
         if ( sass_fn_import?.sassCompilerOpts ) {
 
-            this.compiler.args.sass = sass_fn_import.sassCompilerOpts(
-                this.compiler.args.sass,
-            );
+            this.compiler.args.sass = this.compiler.parseArgs(
+                this.compiler.args,
+                {
+                    sass: sass_fn_import.sassCompilerOpts(
+                        {
+                            config: this.config,
+                            console: this.console,
+                            params: this.params,
+                        },
+                        this.compiler.args.sass,
+                    ),
+                },
+            ).sass;
         }
 
         // runs templates sub-stage and returns