+++ src/ts/functions/sassCompilerOpts.ts
@@ -10,7 +10,15 @@
 
 import type * as sass from "sass-embedded";
 
-import type { Stage } from '@maddimathon/build-utilities';
+import type {
+    CLI,
+    Config,
+    Stage,
+} from '@maddimathon/build-utilities';
+
+import type {
+    Logger,
+} from '@maddimathon/build-utilities/internal';
 
 import { sassCompilerFunctions } from './sassCompilerFunctions.js';
 
@@ -24,6 +32,11 @@ import { sassCompilerFunctions } from './sassCompilerFunctions.js';
 export function sassCompilerOpts<
     T_Partial extends Stage.Compiler.Args.Sass | Partial<Stage.Compiler.Args.Sass>
 >(
+    args: {
+        config: Config.Class,
+        console: Logger,
+        params: CLI.Params,
+    },
     partial?: T_Partial,
 ): T_Partial & Partial<Stage.Compiler.Args.Sass> {
 
@@ -39,7 +52,7 @@ export function sassCompilerOpts<
 
         functions: {
             ...partial?.functions ?? {},
-            ...sassCompilerFunctions(),
+            ...sassCompilerFunctions( args ),
         },
     } satisfies T_Partial & Partial<Stage.Compiler.Args.Sass>;
 }